Frontend Logic Backbone
Renamed grid_+_algorithm_search.py to grid_and_algorithm_search.py for easier access. Made sure it was up to date with the colab from google that we had.

Also added in the backbone for Frontend Logic to run the simulation and have proper UI fitting that.

Created an assets folder to also include any images used to help generate the visuals.